Create PostClient class with GET method
