+++ src/post/components/PostCardsList/PostCardsList.tsx
@@ -1,4 +1,4 @@
-import { Post } from "../../client/types";
+import { Post } from "../../types";
 import PostCard from "../PostCard/PostCard";
 import "./PostCardsList.css";
