+++ src/post/types.ts
@@ -1,0 +1,9 @@
+export interface Post {
+  id: number;
+  title: string;
+  date: Date;
+  author: string;
+  content: string;
+  imageUrl: string;
+  altImageText: string;
+}
